pyproject.toml
[project]
name = "pipesim"
version = "0.1.0"
description = "Data pipeline simulator with retries, backpressure, idempotency, and replay"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["PyYAML>=6.0"]

[project.scripts]
pipesim = "pipesim.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
